Add support for custom rules
