+++ packages/earljs/src/internals.ts
@@ -1,5 +1,5 @@
 export { Control, ValidationResult } from './Control'
 export { Expectation, ExpectationOptions, getControl } from './Expectation'
+export { buildSmartEqResult, ExpectedEqual, loadSmartEqRules, SmartEqResult, SmartEqRules } from './isEqual/rules'
 export * from './matchers'
 export * from './plugins'
-export * from './validators/smartEq'
