+++ packages/earljs/src/isEqual/isEqualUnknown.ts
@@ -2,6 +2,7 @@ import { Matcher } from '../matchers'
 import { EqualityOptions } from './EqualityOptions'
 import { isEqualNumber } from './isEqualNumber'
 import { isEqualObject } from './isEqualObject'
+import { smartEqRules } from './rules'
 
 export function isEqualUnknown(
   value: unknown,
@@ -12,6 +13,13 @@ export function isEqualUnknown(
 ) {
   if (other instanceof Matcher) {
     return other.check(value)
+  }
+
+  for (const rule of smartEqRules) {
+    const ruleResult = rule(value, other, options.ignorePrototypes)
+    if (ruleResult) {
+      return ruleResult.result === 'success'
+    }
   }
 
   if (typeof value !== typeof other) {
