+++ packages/earljs/src/plugins/types.ts
@@ -1,5 +1,5 @@
 import { Expectation } from '../Expectation'
-import { SmartEqResult } from '../validators/smartEq'
+import { SmartEqResult } from '../isEqual/rules'
 
 export type DynamicValidator<T> = (this: Expectation<T>, ...args: any[]) => void | Promise<void>
 export type DynamicMatcher = (...args: any[]) => any // real type should be Matcher but can be cast to anything for improved DX
